Use env-var for auth-helper-url on github action.
There is no need for this to be a secret. In fact it's
not a secret. But, I didn't see a way to add non-secret
environment variables. The docs say add them to the action
script but that's entirely beside the point. The point is
every user wants a difference setting but the action is
checked into the repo so putting them there would not work.